refactor(live): split LiveDO into ConnectionDO + TopicDO
Replace the one-class, two-role LiveDO with two cohesive Durable Object
classes sharing live-protocol.ts (ADR 0025):

- ConnectionDO — owns one client's SSE stream, subscription list, and the
  persisted generation. Handles /connect /subscribe /unsubscribe /deliver
  /probe. Persists only the generation counter; no subscribers table.
- TopicDO — owns the durable subscribers registry, the publish fan-out, and
  the alarm reap. Handles /register /deregister /publish. The subscribers
  table + idempotent misses-column guard live only in its constructor.

Cross-role misroutes are unrepresentable at the type level: each role
resolves its counterpart only through the binding typed to the counterpart
class (TopicDO via CONNECTION_DO, ConnectionDO via TOPIC_DO). There is no
binding through which a publish reaches a connection or a deliver reaches a
topic.

CONNECTION_DO + TOPIC_DO bindings replace LIVE_DO; a v2 wrangler migration
adds both sqlite classes and deleted_classes: ["LiveDO"]. Live-view protocol
is byte-for-byte unchanged: generation-based stale detection,
reachable-mismatch-only pruning, the consecutive-miss reap, the 60s alarm
cadence, and the 2s fan-out timeout all carry over.

Behavior-preserving structural split. fate-live + fate-seam (18/18), the full
vitest suite (229), and 22-pano-live e2e all green with assertions unchanged.
