+++ apps/web/wrangler.jsonc
@@ -27,13 +27,24 @@
 			"migrations_dir": "worker/db/drizzle/migrations"
 		}
 	],
-	// LiveDO — the one Durable Object in phoenix (ADR 0023): cross-isolate live
-	// view fan-out over SSE. Connection-role instances own client SSE streams;
-	// topic-role instances hold durable subscriber registries in SQLite storage.
+	// The two live-fan-out Durable Objects (ADR 0023, split per ADR 0025):
+	// `ConnectionDO` instances own client SSE streams + subscription lists +
+	// generation; `TopicDO` instances hold durable subscriber registries in
+	// SQLite storage and drive the publish fan-out + alarm reap.
 	"durable_objects": {
-		"bindings": [{ "name": "LIVE_DO", "class_name": "LiveDO" }]
+		"bindings": [
+			{ "name": "CONNECTION_DO", "class_name": "ConnectionDO" },
+			{ "name": "TOPIC_DO", "class_name": "TopicDO" }
+		]
 	},
-	"migrations": [{ "tag": "v1", "new_sqlite_classes": ["LiveDO"] }],
+	"migrations": [
+		{ "tag": "v1", "new_sqlite_classes": ["LiveDO"] },
+		{
+			"tag": "v2",
+			"new_sqlite_classes": ["ConnectionDO", "TopicDO"],
+			"deleted_classes": ["LiveDO"]
+		}
+	],
 	"rules": [{ "type": "Text", "globs": ["**/*.sql"], "fallthrough": true }],
 	"vars": {
 		"ENVIRONMENT": "development"
